refactor: rename directory `colorful` to `color`
