+++ packages/@core/base/shared/src/color/generator.ts
