+++ packages/utils/src/index.ts
@@ -1,3 +1,3 @@
 export * from './helpers';
-export * from '@vben-core/shared/colorful';
+export * from '@vben-core/shared/color';
 export * from '@vben-core/shared/utils';
